SideBarのコンポーネント化、inserfaceを使用して型定義
